Possibility to bake curves & surfaces.
